// apps/backend/src/modules/recipes/enums/recipe.enums.ts
export enum RecipeStatus {
  DRAFT = 'draft',
  PRIVATE = 'private',
  PENDING = 'pending',
  PUBLIC = 'public',
  REJECTED = 'rejected',
}

export enum RecipeType {
  PERSONAL = 'personal',
  COMMUNITY = 'community',
}

export enum Difficulty {
  EASY = 'easy',
  MEDIUM = 'medium',
  HARD = 'hard',
}